# docs/openapi.yaml
openapi: 3.0.1
info:
  title: 'EventXperience Backend API'
  description: 'Generated API reference for the EventXperience backend. Responses use the shared ApiResponse envelope unless noted otherwise.'
  version: v1
paths:
...
  /api/events:
    get:
      tags:
        - events
      operationId: Events_GetEvents
      parameters:
        - name: search
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: isPrivate
          in: query
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: category
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: tags
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: city
          in: query
          schema:
            '$ref': '#/components/schemas/String7'
        - name: lat
          in: query
          schema:
            '$ref': '#/components/schemas/Double4'
        - name: lng
          in: query
          schema:
            '$ref': '#/components/schemas/Double4'
        - name: radiusKm
          in: query
          schema:
            '$ref': '#/components/schemas/Double4'
        - name: sortBy
          in: query
          schema:
            type: integer
            default: 0
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfPagedResponseOfEventResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
...
  '/api/events/{eventId}/register':
    post:
      tags:
        - events
      operationId: EventRegistration_Register
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int329'
      responses:
        201:
          description: Created
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    delete:
      tags:
        - events
      operationId: EventRegistration_Unregister
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int329'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/MessageResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/events/{eventId}/registrations':
    get:
      tags:
        - events
      operationId: EventRegistration_GetRegistrations
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int329'
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventRegistrationResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventRegistrationResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventRegistrationResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
  '/api/events/{eventId}/registrations/me':
    get:
      tags:
        - events
      operationId: EventRegistration_CheckRegistration
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int329'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  /api/events/batch/register:
    post:
      tags:
        - events
      operationId: EventRegistration_BatchRegister
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/BatchRegistrationRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/BatchRegistrationRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/BatchRegistrationRequest'
        required: true
      responses:
        207:
          description: Multi-Status
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfBatchRegistrationResultResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfBatchRegistrationResultResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfBatchRegistrationResultResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
    delete:
      tags:
        - events
      operationId: EventRegistration_BatchUnregister
      requestBody:
        content:
          application/json:
            schema:
              '$ref': '#/components/schemas/BatchRegistrationRequest'
          text/json:
            schema:
              '$ref': '#/components/schemas/BatchRegistrationRequest'
          'application/*+json':
            schema:
              '$ref': '#/components/schemas/BatchRegistrationRequest'
        required: true
      responses:
        207:
          description: Multi-Status
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfBatchRegistrationResultResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfBatchRegistrationResultResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfBatchRegistrationResultResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
  '/api/users/{userId}/events/registered':
    get:
      tags:
        - users
      operationId: UserEventRegistration_GetRegisteredEvents
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            '$ref': '#/components/schemas/Int329'
        - name: page
          in: query
          schema:
            '$ref': '#/components/schemas/Int325'
        - name: pageSize
          in: query
          schema:
            '$ref': '#/components/schemas/Int326'
      responses:
        200:
          description: OK
          content:
            text/plain:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventRegistrationResponse'
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventRegistrationResponse'
            text/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfIEnumerableOfEventRegistrationResponse'
        400:
          description: 'Bad request, validation failure, or CSRF failure.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        500:
          description: 'Unexpected server error.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        401:
          description: 'Authentication is required or the access token is invalid/expired.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
        403:
          description: 'The authenticated user does not have permission to perform this action.'
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/ApiResponseOfObject'
      security:
        - bearerAuth:
            []
...
components:
  schemas:
...
    ApiError:
      required:
        - code
      type: object
      properties:
        code:
          '$ref': '#/components/schemas/String'
        details:
          default: null
      nullable: true
...
    ApiResponseOfBatchRegistrationResultResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/BatchRegistrationResultResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfIEnumerableOfEventRegistrationResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/IEnumerableOfEventRegistrationResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfObject:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    ApiResponseOfPagedResponseOfEventResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/PagedResponseOfEventResponse'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    BatchRegistrationFailure:
      type: object
      properties:
        eventId:
          '$ref': '#/components/schemas/Int32'
        reason:
          '$ref': '#/components/schemas/String'
    BatchRegistrationRequest:
      required:
        - eventIds
      type: object
      properties:
        eventIds:
          '$ref': '#/components/schemas/ListOfInt32'
    BatchRegistrationResultResponse:
      type: object
      properties:
        succeeded:
          '$ref': '#/components/schemas/ListOfInt32'
        failed:
          '$ref': '#/components/schemas/ListOfBatchRegistrationFailure'
      nullable: true
...
    Boolean:
      type: boolean
...
    DateTime:
      type: string
      format: date-time
      nullable: true
    DateTime2:
      type: string
      format: date-time
...
    Double:
      type: number
      format: double
      nullable: true
...
    Double4:
      type: number
      format: double
      default: null
...
    EventCategory:
      type: integer
...
    EventHostClubResponse:
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        name:
          '$ref': '#/components/schemas/String'
        description:
          '$ref': '#/components/schemas/String'
        clubType:
          '$ref': '#/components/schemas/String'
        clubImage:
          '$ref': '#/components/schemas/String'
        memberCount:
          '$ref': '#/components/schemas/Int32'
        eventCount:
          '$ref': '#/components/schemas/Int32'
        availableEventCount:
          '$ref': '#/components/schemas/Int32'
        isPrivate:
          '$ref': '#/components/schemas/Boolean'
        email:
          '$ref': '#/components/schemas/String2'
        phone:
          '$ref': '#/components/schemas/String2'
        rating:
          '$ref': '#/components/schemas/Double'
        websiteUrl:
          '$ref': '#/components/schemas/String2'
        location:
          '$ref': '#/components/schemas/String2'
      nullable: true
...
    EventLifecycleState:
      type: integer
    EventRegistrationResponse:
      required:
        - id
        - userId
        - eventId
        - createdAt
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        userId:
          '$ref': '#/components/schemas/Int32'
        eventId:
          '$ref': '#/components/schemas/Int32'
        createdAt:
          '$ref': '#/components/schemas/DateTime2'
...
    EventResponse2:
      type: object
      properties:
        id:
          '$ref': '#/components/schemas/Int32'
        name:
          '$ref': '#/components/schemas/String'
        description:
          '$ref': '#/components/schemas/String'
        location:
          '$ref': '#/components/schemas/String'
        imageUrls:
          '$ref': '#/components/schemas/ListOfString'
        isPrivate:
          '$ref': '#/components/schemas/Boolean'
        maxParticipants:
          '$ref': '#/components/schemas/Int32'
        registerCost:
          '$ref': '#/components/schemas/Int32'
        startTime:
          '$ref': '#/components/schemas/DateTime2'
        endTime:
          '$ref': '#/components/schemas/DateTime'
        clubId:
          '$ref': '#/components/schemas/Int32'
        currentVersionNumber:
          '$ref': '#/components/schemas/Int32'
        createdAt:
          '$ref': '#/components/schemas/DateTime2'
        lifecycleState:
          '$ref': '#/components/schemas/EventLifecycleState'
        status:
          '$ref': '#/components/schemas/EventStatus'
        category:
          '$ref': '#/components/schemas/EventCategory'
        venueName:
          '$ref': '#/components/schemas/String2'
        city:
          '$ref': '#/components/schemas/String2'
        latitude:
          '$ref': '#/components/schemas/Double'
        longitude:
          '$ref': '#/components/schemas/Double'
        tags:
          '$ref': '#/components/schemas/ListOfString'
        registrationCount:
          '$ref': '#/components/schemas/Int32'
        distanceKm:
          '$ref': '#/components/schemas/Double'
        club:
          '$ref': '#/components/schemas/EventHostClubResponse'
...
    EventStatus:
      type: integer
...
    IEnumerableOfEventRegistrationResponse:
      type: array
      items:
        '$ref': '#/components/schemas/EventRegistrationResponse'
      nullable: true
    IEnumerableOfEventResponse:
      type: array
      items:
        '$ref': '#/components/schemas/EventResponse2'
...
    Int32:
      type: integer
      format: int32
...
    Int325:
      type: integer
      format: int32
      default: 1
    Int326:
      type: integer
      format: int32
      default: 20
...
    Int329:
      maximum: 2147483647
      minimum: 1
      type: integer
      format: int32
...
    ListOfBatchRegistrationFailure:
      type: array
      items:
        '$ref': '#/components/schemas/BatchRegistrationFailure'
...
    ListOfInt32:
      type: array
      items:
        '$ref': '#/components/schemas/Int32'
...
    ListOfString:
      type: array
      items:
        '$ref': '#/components/schemas/String'
...
    MessageResponse:
      type: object
      properties:
        success:
          '$ref': '#/components/schemas/Boolean'
        message:
          '$ref': '#/components/schemas/String'
        data:
          '$ref': '#/components/schemas/Object'
        error:
          '$ref': '#/components/schemas/ApiError'
        meta:
          '$ref': '#/components/schemas/Object'
...
    Object: {}
...
    PagedResponseOfEventResponse:
      required:
        - items
        - totalCount
        - page
        - pageSize
      type: object
      properties:
        items:
          '$ref': '#/components/schemas/IEnumerableOfEventResponse'
        totalCount:
          '$ref': '#/components/schemas/Int32'
        page:
          '$ref': '#/components/schemas/Int32'
        pageSize:
          '$ref': '#/components/schemas/Int32'
        totalPages:
          '$ref': '#/components/schemas/Int32'
      nullable: true
...
    String:
      type: string
...
    String2:
      type: string
      nullable: true
...
    String7:
      type: string
      default: null
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Send the access token in the Authorization header as `Bearer {token}` for protected API routes.'
      scheme: bearer
      bearerFormat: JWT
...